Atualizações do Desafio

+++ src/database/migrations/1587052577410-CreateTransactionTable.ts
@@ -25,7 +25,7 @@ export default class CreateTransactionTable1587052577410
           },
           {
             name: 'value',
-            type: 'double',
+            type: 'decimal',
             isNullable: false,
           },
           {
